Use CatchAll observer within DefaultDnsClient to centralize exception handling
